==> code/Code_Bazzigher/S2_Mata_et_al._2016_Replication.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
---

```{r}
rm(list = ls())
```

# Library
```{r}
library(tidyverse)
library(readr)
library(dplyr)
library(lme4)
library(lmerTest)
```

# Load data
##### We first prepare World Values Survey (WVS) data 
##### We (e.g., exclude cases that have missing information on key variables, such as age, sex, 
##### education, etc.; dichotomize variables; transform risk taking item)
```{r}
dateipfad <- "\\\\kt.ktzh.ch\\Personal\\Home01\\B117T23\\Documents\\Laura Bazzigher\\Privat\\Studium\\Riskktaking\\Data"

data_wv5 <- readRDS(file.path(dateipfad, "F00007944-WV5_Data_R_v20180912.rds"))

load(file.path(dateipfad, "WV6_Data_R_v20201117.rdata"))
head(data_wv5)
head(WV6_Data_R_v20201117)
```
# Load Country names
#####  read in a file with the country names and respective numeric codes used in WVS
```{r}
countrynames <- read.csv(file.path(dateipfad, "countrynames.txt"), header = FALSE, as.is = TRUE) 
```

# Rename WVS5 
##### select a subset of the WVS data (Wave 5), including wave, country, 10 Schwartz items (including the risk item), and demographics
```{r} 
data5 <- data_wv5[, c("V1", "V2", "V80","V81","V82","V83","V84","V85","V86","V87","V88","V89", "V235", "V237", "V238", "V241", "V55", "V56", "V260")]
names(data5) <- c("wave", "country","creativity","money","security","goodtime","help","success","risk","proper","environment","tradition", "sex", "age", "education", "employed", "married", "children", "year")
head(data5)
```


# DATA CHECK
##### comparing the N from data5 against the counts from counts provided in www.worldvaluessurvey.org
```{r}
data5[,"country_lab"]=NA
countries=unique(data5$country)
for(row in 1:length(countries))
{
  data5[data5$country==countries[row],"country_lab"]=countrynames[countrynames[,1]==countries[row],2]
}
table(data5$country_lab)
```

# Rename WVS6 
##### select a subset of the WVS data (Wave 6), including wave, country, 10 Schwartz items (including the risk item), and demographics 
```{r}
data6 = WV6_Data_R_v20201117[, c("V1", "V2","V70","V71","V72","V73","V74B","V75","V76","V77","V78","V79", "V240", "V242", "V248", "V229", "V57", "V58", "V262")]
names(data6) = c("wave", "country","creativity","money","security","goodtime","help","success","risk","proper","environment","tradition", "sex", "age", "education", "employed", "married", "children", "year")

head(data6)
```
# DATA CHECK
```{r}
data6[,"country_lab"]=NA
countries=unique(data6$country)
for(row in 1:length(countries))
{
  data6[data6$country==countries[row],"country_lab"]=countrynames[countrynames[,1]==countries[row],2] 
}
table(data6$country_lab)
```


# Bind data 
##### Data looks ok so we bind the two waves...
```{r}
data = rbind(data5, data6)
head(data)
```
# Laura: Check Country and Country-nr
```{r}
unique_countries <- unique(data[, c("country_lab", "country")])
unique_countries
```


# PARTICIPANT EXCLUSION 
#### We exclude individuals for which we do not know sex, age, education, employment status, marital status, or number of children
```{r}
data = subset(data, risk > 0 & sex > 0 & age >0 & education > 0 & employed > 0 & married > 0 & children >= 0)

length(unique(data$country)) # number of countries
nrow(data) # number of individuals
range(data$age) # age
table(data$sex) # sex table(data$sex)/nrow(data)
```

# Age category 
##### create a categorical age variable (for example, to plot response frequencies by category)
```{r}
data$agecat[data$age<20]="15-19"
data$agecat[data$age>=20 & data$age <30] = "20-29"
data$agecat[data$age>=30 & data$age <40] = "30-39"
data$agecat[data$age>=40 & data$age <50] = "40-49"
data$agecat[data$age>=50 & data$age <60] = "50-59"
data$agecat[data$age>=60 & data$age <70] = "60-69"
data$agecat[data$age>=70 & data$age <80] = "70-79"
data$agecat[data$age>=80] = "80+"

table(data$age)
table(data$agecat)
```
#EDUCATION
#1 No formal education
#2 Incomplete primary school
#3 Complete primary school
#4 Incomplete secondary school: technical/vocational type
#5 Complete secondary school: technical/vocational type
#6 Incomplete secondary: university-preparatory type
#7 Complete secondary: university-preparatory type
#8 Some university-level education, without degree
#9 University-level education, with degree


# education category
##### create a categorical education variable (with fewer categories than the original)
```{r}
attach(data)
data$education_cat[education < 3] = "incomplete or no primary education"
data$education_cat[education > 2 & education <= 6] <- "no uni"
data$education_cat[education >= 7] <- "uni"
detach(data)
table(data$education)
```
#EMPLOYED
#Yes, has paid employment:
#Full time employee (30 hours a week or more) 1 
#Part time employee (less than 30 hours a week) 2 
#Self employed 3
#No, no paid employment:
#Retired/pensioned 4 
#Housewife not otherwise employed 5
#Student 6 
#Unemployed 7
#Other (write in): 8


# Dichotomizing Variables
##### This helps estimating and interpreting the models later on...
```{r}
data$sex = ifelse(data$sex == 1, 0, 1) # sex: male vs. female
data$children = ifelse(data$children == 0, 0, 1) # children: no vs. yes
data$married = ifelse(data$married == 1, 1, 0) # married: yes vs. no
data$employed = ifelse(data$employed < 4, 1, 0) # employed: yes vs. no
data$education = ifelse(data$education < 4, 0, 1) # education: no primary vs. primary+ 
```

# Hardship (Countryfacts) 
##### Read in country names, assign country labels to WVS data, and add the hardship indicators + hardship index 
#### (see below for more analyses on the hardship index)

#####  read in file that contains hardship indicators manually collected from CIA factbook, WHO, and World Bank 
#### (see Supplemental Materials for URL sources)
```{r}
library(readxl)

excel_path <- "\\\\kt.ktzh.ch\\Personal\\Home01\\B117T23\\Documents\\Laura Bazzigher\\Privat\\Studium\\Riskktaking\\Data\\Hardship_complete_2024.xlsx"
countryfacts <- read_excel(excel_path)

labels <- c("code", "country", "codeWVS", "Homicide", "GDP", "InfMort", "LifeExp", "GINI", "GenderPEdu")
names(countryfacts) <- labels

unique(data$country) %in% countryfacts$codeWVS # check that all countries in the subset of the WVS data are included in the countryfacts file

countryfacts
```

# CHECK New code Laura
##### Checking Effectiveness BEFORE Imputation Method (to compare with new code)
```{r}
summary(countryfacts[, c("Homicide", "GDP", "InfMort", "LifeExp", "GINI", "GenderPEdu")])
```

# log transform
```{r}
countryfacts$Homicide=log(countryfacts$Homicide)
countryfacts$GDP=log(countryfacts$GDP)
countryfacts$InfMort=log(countryfacts$InfMort)
countryfacts$LifeExp=log(countryfacts$LifeExp)
countryfacts
```

# CHECK New code Laura
##### Checking Effectiveness AFTER Imputation Method (to compare with new code)
```{r}
summary(countryfacts[, c("Homicide", "GDP", "InfMort", "LifeExp", "GINI", "GenderPEdu")])
```


# scale and reverse
```{r}
countryfacts$Homicide=scale(countryfacts$Homicide)
countryfacts$GDP=scale(-countryfacts$GDP)
countryfacts$InfMort=scale(countryfacts$InfMort)
countryfacts$LifeExp=scale(-countryfacts$LifeExp)
countryfacts$GINI=scale(countryfacts$GINI)
countryfacts$GenderPEdu=scale(-countryfacts$GenderPEdu)
countryfacts

#panel.hist(countryfacts[,4:9]) # plot histogram of country indicators
#dev.print(postscript,"histindicatorstrans.eps",width=6, height=6,horizontal=FALSE,onefile=FALSE)
```

# Convert into numerical variables
#### not in the original code 
```{r}
countryfacts$Homicide <- as.numeric(countryfacts$Homicide)
countryfacts$GDP <- as.numeric(countryfacts$GDP)
countryfacts$InfMort <- as.numeric(countryfacts$InfMort)
countryfacts$LifeExp <- as.numeric(countryfacts$LifeExp)
countryfacts$GINI <- as.numeric(countryfacts$GINI)
countryfacts$GenderPEdu <- as.numeric(countryfacts$GenderPEdu)
countryfacts
```

# This code does not work. I do not understand why, all the variables are numeric but it always says, that they are not. Since out counrtyfacts-document has no missing values, I do not run it.```{r}
# IMPUTE hardship indicators w/ median
#```{r}
#for (counter in 5:9)
#{
#  countryfacts[is.na(countryfacts[,counter]),counter]=median(countryfacts[,counter],na.rm=TRUE)
#}
#```

# Impute hardship indicators w/ median
```{r}
# IMPUTE hardship indicators w/ median
for (counter in 4:9) {
  # Convert the column to numeric (in case it's not already numeric)
  countryfacts[[counter]] <- as.numeric(as.character(countryfacts[[counter]]))
  
  # Replace non-numeric values with NA
  countryfacts[[counter]][!is.na(countryfacts[[counter]]) & !is.numeric(countryfacts[[counter]])] <- NA
  
  # Impute missing values with the median
  countryfacts[[counter]][is.na(countryfacts[[counter]])] <- median(countryfacts[[counter]], na.rm = TRUE)
}
countryfacts
```


# Create a Hardship Index 
```{r}
countryfacts$hardship = (countryfacts$Homicide + countryfacts$GDP + countryfacts$InfMort + countryfacts$LifeExp + countryfacts$GINI + countryfacts$GenderPEdu) / 6
countryfacts
```

# SUP MATERIALS:Correlation between hardship indicators 

## Plot scatter plot matrix
```{r}
panel.cor = function(x, y, digits = 2, cex.cor, ...)
{
  usr = par("usr"); on.exit(par(usr))
  par(usr = c(0, 1, 0, 1))
  # correlation coefficient
  r = cor(x, y,use="complete.obs")
  txt = format(c(r, 0.123456789), digits = digits)[1]
  txt = paste("r= ", txt, sep = "")
  text(0.5, 0.6, txt)
  
  # p-value calculation
  p = cor.test(x, y,use="complete.obs")$p.value
  txt2 = format(c(p, 0.123456789), digits = digits)[1]
  txt2 = paste("p= ", txt2, sep = "")
  if(p<0.01) txt2 = paste("p ", "<0.01", sep = "")
  text(0.5, 0.4, txt2)
}
pairs(countryfacts[,4:10], upper.panel = panel.cor,las=1,cex.labels=.9) 
dev.print(postscript,"scatter_indicators.eps",width=8, height=8,horizontal=FALSE,onefile=FALSE)
```

## Cronbach's alpha for Hardship Indicators
```{r}
library(psych)
alpha(countryfacts[,4:9]) # calculate consistency (Cronbach's alpha) for the hardship indicators
```

# Add country label and hardship 
##### add the country label and hardship variables to the data matrix (i.e., data)
```{r}
data[,"country_lab"]=NA
data[,"hardship"]=NA
countries=unique(data$country)
for(row in 1:length(countries))
{
  data[data$country==countries[row],"country_lab"]=countryfacts[countryfacts$codeWVS==countries[row],"code"]
  
  if(!is.na(countryfacts[countryfacts$codeWVS==countries[row],"hardship"]))
  {
    data[data$country==countries[row],"hardship"]=rep(countryfacts[countryfacts$codeWVS==countries[row],"hardship"],length(sum(data$country==countries[row])))
  }
}
```

# Transformation Risk item 
##### Transfrom risk item such that high values represent more risk taking
```{r}
data$risk = 6 - data$risk + 1
```

# Risk item as ordinal variable 
```{r}
data$risk_ord=data$risk 
```

# Transform risk variable into T-score 
##### (mean = 50, sd = 10)
```{r}
data$risk = 10*scale(data$risk, center=TRUE,scale=TRUE)+50
head(data)
```


# CHECK New code Laura 
##### Checking Log Transformation and Scaling/Reversing (to compare with new code)
```{r}
library(ggplot2)

summary(countryfacts[, c("Homicide", "GDP", "InfMort", "LifeExp")])
# Visualize distributions before and after transformation
ggplot(countryfacts, aes(x = GDP)) + 
  geom_histogram(bins = 20, fill = "lightblue", color = "black") +
  ggtitle("Distribution of GDP Before Transformation") +
  xlab("GDP") + ylab("Frequency")

ggplot(countryfacts, aes(x = scale(-GDP))) + 
  geom_histogram(bins = 20, fill = "lightblue", color = "black") +
  ggtitle("Distribution of GDP After Scaling/Reversing") +
  xlab("Scaled/Reversed GDP") + ylab("Frequency")
```


# CHECK New code Laura
##### Checking Effectiveness of Imputation Method (to compare with new code)
```{r}
summary(countryfacts[, c("Homicide", "GDP", "InfMort", "LifeExp", "GINI", "GenderPEdu")])
```

# MIXED-EFFECTS LINEAR MODELING 


# model0
#### intercept only model
```{r}
# intercept only model
model0 = lmer(risk ~ 1 + (1|country),data = data)
summary_model0=summary(model0)
summary_model0
```

# model1
#### age, gender 
```{r}
# age, sex 
model1 = lmer(risk ~ 1 +scale(age)+factor(sex) + (1+scale(age)+factor(sex)|country),data = data, control = lmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))

summary_model1=summary(model1)
summary_model1
```


# model2
#### age, gender, and covariates (children, marital status, employement status, education)
```{r}
#model 2

options(scipen = 100)
model_2 <- lmer(risk ~ 1 + scale(age) + factor(sex) + factor(children) + 
                 factor(married) + factor(education) + factor(employed) +
                 (1 + scale(age) + factor(sex) + factor(children) + factor(employed) +
                  factor(married) | country),
                data = data, control = lmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))


summary_model_2 = summary(model_2)
summary_model_2
```


# model3
#### age, gender, and covariates (children, marital status, employement status, education) and HARDSHIP
```{r}
model3 <- lmer(risk ~ 1+scale(age)*hardship+factor(sex)*hardship + factor(children) + factor(married) + factor(employed) + factor(education)+ (1+scale(age)+factor(sex)+ factor(children) + factor(married) + factor(employed) + factor(education)|country),data = data,control=lmerControl(optCtrl=list(maxfun=100000),optimizer="bobyqa"),REML = FALSE)

summary_model3 = summary(model3) 
summary_model3
```




# ANOVA
```{r}
anova(model0,model1)
anova(model1,model_2)
anova(model_2,model3)
```

```{r}
coefsallmodels=rbind(summary_model1$coefficients,
summary_model_2$coefficients,
summary_model3$coefficients[c(1:2,4:8,3,9:10),])

write.csv(coefsallmodels,"coefsallmodels.csv")
```


```{r}
# Extract random effects for 'country'
# Assuming 'model' is your fitted lmer model
random_effects <- ranef(model_2)
random_effects
```


```{r}
regression_results_WVS <- data %>%
  group_by(country) %>%
  do(model = lm(risk ~ scale(age) + sex, data = .)) %>%
  summarize(
    country = first(country),
    intercept_regression = coef(summary(model))[1, 1],
    slope_age_regression = coef(summary(model))[2, 1],
    slope_gender_regression = coef(summary(model))[3, 1]
  )

regression_results_WVS

random_effects <- ranef(model_2)
random_effects

colnames(random_effects)
```

```{r}
coefsallmodels=rbind(summary_model1$coefficients,
summary_model_2$coefficients,
summary_model3$coefficients[c(1:2,4:8,3,9:10),])

write.csv(coefsallmodels,"coefsallmodels.csv")
```

```{r}
# Predict risk-taking propensity using model_2
data$predicted_risk <- predict(model_2, type = "response")
```

```{r}
# Create density plots for all countries combined
ggplot(data, aes(x = scale(age), fill = factor(sex))) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("blue", "red")) +
  labs(x = "Age (scaled)", y = "Density", fill = "Sex") +
  theme_minimal() +
  theme(legend.position = "top")
```


```{r}
library(ggplot2)
library(dplyr)

# Stelle sicher, dass 'sex' als Faktor kodiert ist
data$sex <- factor(data$sex)

# Plotte die Daten
plot <- ggplot(data, aes(x = age)) + 
  geom_density(aes(fill = sex), alpha = 0.5) + 
  scale_fill_manual(values = c("blue", "red")) +
  facet_wrap(~ country_lab, scales = "free_y", ncol = 4) + 
  geom_smooth(aes(x = age, y = risk, color = sex, group = sex), method = "lm", linetype = "dashed", se = FALSE) +
  theme_minimal() +
  labs(x = "Age", y = "Propensity for Risk Taking (T score)", fill = "Sex", color = "Sex") +
  theme(legend.position = "top") +
  scale_y_continuous(limits = c(30, NA)) # Setze das untere Limit auf 30 und das obere auf NA, um es automatisch zu bestimmen

print(plot)

ggsave("myplot.png", plot, width = 10, height = 30, units = "in", dpi = 300)
```

```{r}
library(ggplot2)
library(dplyr)

# Stelle sicher, dass 'sex' als Faktor kodiert ist
data$sex <- factor(data$sex)

# Plotte die Daten
plot <- ggplot(data, aes(x = age)) + 
  geom_density(aes(fill = sex), alpha = 0.5) + 
  scale_fill_manual(values = c("blue", "red")) +
  facet_wrap(~ country_lab, scales = "free_y", ncol = 4) + 
  geom_smooth(aes(x = age, y = risk, color = sex, group = sex), method = "lm", linetype = "dashed", se = FALSE) +
  theme_minimal() +
  labs(x = "Age", y = "Propensity for Risk Taking (T score)", fill = "Sex", color = "Sex") +
  theme(legend.position = "top") +
  coord_cartesian(ylim = c(30, NA)) # Setze das untere Limit auf 30 und das obere auf NA, um es automatisch zu bestimmen

print(plot)

ggsave("myplot.png", plot, width = 10, height = 30, units = "in", dpi = 300)
```

```{r}
library(ggplot2)
library(dplyr)

# Stelle sicher, dass 'sex' als Faktor kodiert ist
data$sex <- factor(data$sex)

# Plotte die Daten für alle Länder zusammen
plot <- ggplot(data, aes(x = age)) + 
  geom_density(aes(fill = sex), alpha = 0.5) + 
  scale_fill_manual(values = c("blue", "red")) +
  geom_smooth(aes(x = age, y = risk, color = sex, group = sex), method = "lm", linetype = "dashed", se = FALSE) +
  theme_minimal() +
  labs(x = "Age", y = "Propensity for Risk Taking (T score)", fill = "Sex", color = "Sex") +
  theme(legend.position = "top") +
  scale_y_continuous(limits = c(30, NA)) # Behalte das untere Limit von 30 bei

print(plot)

ggsave("myplot_all_countries.png", plot, width = 10, height = 8, units = "in", dpi = 300)
```
